Move manager to separate package.
